remove bufcurl.Resolver; everything can just use protoencoding.Resolver
